Rewrite the backend message parser
The owned variant of the old parser incurred a lot of allocation
overhead, particularly for row data messages. Rewrite it so that it
behaves like the borrowed parser, except that the message can optionally
own any dynamically sized data it contains.
