+++ Cargo.toml
@@ -13,3 +13,4 @@ byteorder = "0.5"
 fallible-iterator = "0.1"
 hex = "0.2"
 md5 = "0.2"
+memchr = "0.1"
